requirements.txt
mysql-connector-python==9.5.0
numpy==2.4.2
opencv-contrib-python==4.13.0.90
pillow==12.1.0
